:recycle: [Refactor] 경로 수정
