+++ src/types/api.ts
@@ -1,5 +1,5 @@
 import { QuoteData } from "./quote";
-import { MovingRequestDataWithComplete } from "./moving-request";
+import { MovingRequestDataWithComplete } from "./movingRequest";
 import { MoverDetailData } from "./mover";
 
 // API 구조 참고
